requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
scikit-learn
pandas
argon2-cffi
python-jose[cryptography]
psycopg2-binary
pydantic
email-validator
python-multipart
pdfplumber
sentence-transformers
